aula2 items controller
