+++ server/src/routes.ts
@@ -1,10 +1,11 @@
 import express from "express";
-import knex from "./database/connection";
 
 import PointsController from "./controllers/PointsController";
+import ItemsController from "./controllers/ItemsController";
 
 const routes = express.Router();
 const pointsController = new PointsController();
+const itemsController = new ItemsController();
 
 routes.get("/users", (req, res) => {
   return res.json({ message: "NLW01 aula2" });
@@ -12,23 +13,14 @@ routes.get("/users", (req, res) => {
 
 routes.get("/users/:id", (req, res) => {});
 
-routes.post("/users", (req, res) => {
-  //return res.json(res.body)
-});
+routes.post("/users", (req, res) => {});
 
-routes.get("/items", async (req, res) => {
-  const items = await knex("items").select("*");
-
-  const serializedItems = items.map((item) => {
-    return {
-      id: item.id,
-      title: item.title,
-      image_url: `http://localhost:3333/uploads/${item.image}`,
-    };
-  });
-  return res.json(serializedItems);
-});
+routes.get("/items", itemsController.index);
 
 routes.post("/points", pointsController.create);
+routes.get("/points/:id", pointsController.show);
 
 export default routes;
+
+//service pattern
+//repository pattern data mapper
